refactor(tui): apply PR #155 review cleanups
Address committed-PR review findings from the cost-visibility branch.
Cleanup only, no behavior change.

- gauge.go, tiles_gauge.go: extract joinAnnotationParts helper that
  uses lo.Compact + strings.Join so the dim "resets … · projected …"
  annotation no longer needs a 3-arm switch at each call site.
- tiles_gauge.go, detail_sections.go: build gaugeAllowSet via
  lo.SliceToMap instead of a manual loop in two places.
- model.go: drop the boolKeys helper and call SortedStringKeys on the
  typed map directly — the conversion no longer carries its weight.
- detail_sections.go: flatten buildDetailGaugeLines from 5-deep
  nesting using guard-clause inversion (plain gauge by default,
  override with projection variant only when window AND reset are
  both recognized). Preserves exact behavior.
- gauge.go: relocate gaugeWindowDuration from detail_sections.go to
  its natural home next to the other gauge helpers — it's already
  called from tiles_gauge.go and model_input.go.
- detail_trends.go: trim the redundant first sentence of the
  buildDetailTrendsSectionWithHide doc.
